printing test result

+++ test/utils/read_orlib_fl.hpp
@@ -15,7 +15,7 @@
 
 namespace paal {
 
-static const double MULTIPL = 1000000;
+static const long long MULTIPL = 1000000;
 
 inline long long cast(double d) {
     return (long long)(d * MULTIPL) + 1;
